+++ src/fweight.jl
@@ -1,6 +1,6 @@
 """
-    fweightm(X, v)
-    fweightm!(X::AbstractMatrix, v)
+    fweight(X, v)
+    fweight!(X::AbstractMatrix, v)
 Weight each row of a matrix.
 * `X` : Data (n, p).
 * `v` : A weighting vector (n).
@@ -11,14 +11,14 @@ using Jchemo, LinearAlgebra
 
 X = rand(5, 2) 
 w = rand(5) 
-fweightm(X, w)
+fweight(X, w)
 diagm(w) * X
 
-fweightm!(X, w)
+fweight!(X, w)
 X
 ```
 """ 
-function fweightm(X, v)
+function fweight(X, v)
     X = ensure_mat(X)
     n, p = size(X)
     zX = similar(X)
@@ -28,7 +28,7 @@ function fweightm(X, v)
     zX
 end
 
-function fweightm!(X::AbstractMatrix, v)
+function fweight!(X::AbstractMatrix, v)
     n, p = size(X)
     @inbounds for j = 1:p, i = 1:n
         X[i, j] *= v[i]