+++ src/spca.jl
@@ -151,7 +151,7 @@ function spca!(X::Matrix, weights::Weight; kwargs...)
         fcenter!(X, xmeans)
     end
     sqrtw = sqrt.(weights.w)
-    X .= Diagonal(sqrtw) * X
+    fweight!(X, sqrtw)
     t = similar(X, n)
     T = similar(X, n, nlv)
     P = similar(X, p, nlv)